CDN_backend/requirements.txt
Django==4.1.13
djangorestframework==3.12.4
python-dotenv==1.0.1
requests==2.32.3
drf-yasg==1.21.8